Move TastyInterpreter
